+++ CHANGES.md
@@ -4,6 +4,7 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 
+- 241: also delete organizationunits from admin db
 - 164: introduce actions to automatically deploy on tag and PR
 - 192: Add assignments page with organization hierarchy filter, multiselect role filter, and compact card layout
 - 192: Support importing OPEN assignments via CSV